Convert mock world to be as realistic as possible
