+++ game/src/test/resources/test.properties
@@ -1,21 +1,32 @@
 # Game Settings
 name=Void
 port=43594
+dataPath=../data/
 cachePath=../data/cache/
 savePath=../data/saves/
 interfacesPath=../data/definitions/interfaces.yml
 interfaceTypesPath=../data/definitions/interface-types.yml
 objectsPath=../data/spawns/object-spawns.yml
+stairsPath=../data/map/stairs.yml
+musicPath=../data/map/music-tracks.yml
+navGraphPath=../data/map/nav-graph.yml
+areaPath=../data/map/areas.yml
 objectDefinitionsPath=../data/definitions/objects.yml
 npcDefinitionsPath=../data/definitions/npcs.yml
 itemDefinitionsPath=../data/definitions/items.yml
 animationDefinitionsPath=../data/definitions/animations.yml
 graphicDefinitionsPath=../data/definitions/graphics.yml
 containerDefinitionsPath=../data/definitions/containers.yml
+soundDefinitionsPath=../data/definitions/sounds.yml
+midiDefinitionsPath=../data/definitions/midis.yml
+jingleDefinitionsPath=../data/definitions/jingles.yml
+variableDefinitionsPath=../data/definitions/variables.yml
+dropsPath=../data/spawns/drops.yml
 xteaPath=../data/xteas.dat
-clientBuild=667
+revision=634
+loginLimit=5
 scriptModule=game
-homeX=3087
-homeY=3497
+homeX=3221
+homeY=3219
 loginPerTickCap=25
 eventBufferLimit=25
